third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/the-object-element/object-attributes-expected.txt

This is a testharness.js-based test.
FAIL object.contentWindowassert_equals: The contentWindow's name of the object element should be 'o1'. expected "o1" but got "o"
FAIL object.widthassert_equals: The width should be 100px. expected "100px" but got "auto"
FAIL object.heightassert_equals: The height should be 50px. expected "50px" but got "auto"
Harness: the test ran to completion.
